chore(@embark/utils): convert utils/hosts.js to typescript
